+++ package.json
@@ -157,7 +157,7 @@
   },
   "dependencies": {
     "@docsearch/react": "^3.0.0-alpha.37",
-    "@influxdata/clockface": "^3.3.3",
+    "@influxdata/clockface": "^3.4.1",
     "@influxdata/flux-lsp-browser": "^0.8.5",
     "@influxdata/giraffe": "^2.24.3",
     "@influxdata/influxdb-templates": "0.9.0",
