+++ src/flows/components/VersionSidebar.scss
@@ -10,15 +10,6 @@
   line-height: 1.2;
 }
 
-.current-version--text {
-  font-style: italic;
-  font-weight: normal;
-  line-height: 1.6;
-  color: $c-pool;
-  flex: none;
-  margin: 0px 0px;
-}
-
 .published-date--text {
   font-weight: normal;
   font-size: 15.75px;
