+++ src/flows/components/header/index.tsx
@@ -1,38 +1,23 @@
 // Libraries
-import React, {
-  FC,
-  useCallback,
-  useContext,
-  useState,
-  useEffect,
-  createRef,
-  RefObject,
-} from 'react'
-import {useHistory} from 'react-router-dom'
+import React, {FC, useCallback, useContext, useState, useEffect} from 'react'
 import {useSelector} from 'react-redux'
 
 // Contexts
 import {FlowContext} from 'src/flows/context/flow.current'
-import {FlowListContext} from 'src/flows/context/flow.list'
 import {
   VersionPublishContext,
   VersionPublishProvider,
 } from 'src/flows/context/version.publish'
 import {AppSettingProvider} from 'src/shared/contexts/app'
-import {deletePinnedItemByParam} from 'src/shared/contexts/pinneditems'
 
 // Components
 import {
   Page,
   SquareButton,
-  Icon,
   IconFont,
   ComponentColor,
   ComponentStatus,
   ErrorTooltip,
-  Popover,
-  PopoverInteraction,
-  List,
 } from '@influxdata/clockface'
 
 import AutoRefreshButton from 'src/flows/components/header/AutoRefreshButton'
@@ -43,6 +28,7 @@ import SaveState from 'src/flows/components/header/SaveState'
 import PresentationMode from 'src/flows/components/header/PresentationMode'
 import RenamablePageTitle from 'src/pageLayout/components/RenamablePageTitle'
 import {FeatureFlag} from 'src/shared/utils/featureFlag'
+import MenuButton from 'src/flows/components/header/MenuButton'
 
 // Utility
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
@@ -52,7 +38,6 @@ import {
   postNotebooksShare,
 } from 'src/client/notebooksRoutes'
 import {event} from 'src/cloud/utils/reporting'
-import {downloadImage} from 'src/shared/utils/download'
 import {serialize} from 'src/flows/context/flow.list'
 import {updatePinnedItemByParam} from 'src/shared/contexts/pinneditems'
 import {getOrg} from 'src/organizations/selectors'
@@ -61,74 +46,7 @@ import {getOrg} from 'src/organizations/selectors'
 import {RemoteDataState} from 'src/types'
 
 // Constants
-import {
-  DEFAULT_PROJECT_NAME,
-  PROJECT_NAME,
-  PROJECT_NAME_PLURAL,
-} from 'src/flows'
-
-const backgroundColor = '#07070E'
-
-interface MenuItem {
-  type: 'menuitem'
-  title: string
-  onClick: () => void
-  icon: IconFont
-  disabled?: () => boolean
-  testID?: string
-}
-interface Divider {
-  type: 'divider'
-  title: string
-}
-
-type MenuItemType = MenuItem | Divider
-interface ButtonProp {
-  menuItems: MenuItemType[]
-}
-
-const MenuButton: FC<ButtonProp> = ({menuItems}) => {
-  const triggerRef: RefObject<HTMLButtonElement> = createRef()
-  return (
-    <>
-      <SquareButton
-        ref={triggerRef}
-        icon={IconFont.More}
-        testID="flow-menu-button"
-      />
-      <Popover
-        triggerRef={triggerRef}
-        enableDefaultStyles={false}
-        style={{minWidth: 209}}
-        showEvent={PopoverInteraction.Click}
-        hideEvent={PopoverInteraction.Click}
-        contents={onHide => (
-          <List>
-            {menuItems.map(item => {
-              if (item.type === 'divider') {
-                return <List.Divider key={item.title} />
-              }
-              return (
-                <List.Item
-                  key={item.title}
-                  disabled={item?.disabled ? item.disabled() : false}
-                  onClick={() => {
-                    item?.onClick()
-                    onHide()
-                  }}
-                  testID={item?.testID || ''}
-                >
-                  <Icon glyph={item?.icon} />
-                  <span style={{paddingLeft: '10px'}}>{item.title}</span>
-                </List.Item>
-              )
-            })}
-          </List>
-        )}
-      />
-    </>
-  )
-}
+import {DEFAULT_PROJECT_NAME, PROJECT_NAME} from 'src/flows'
 
 interface Share {
   id: string
@@ -136,10 +54,8 @@ interface Share {
 }
 
 const FlowHeader: FC = () => {
-  const {remove, clone} = useContext(FlowListContext)
   const {flow, updateOther} = useContext(FlowContext)
-  const {handlePublish, versions} = useContext(VersionPublishContext)
-  const history = useHistory()
+  const {handlePublish} = useContext(VersionPublishContext)
   const {id: orgID} = useSelector(getOrg)
   const [sharing, setSharing] = useState(false)
   const [share, setShare] = useState<Share>()
@@ -208,86 +124,6 @@ const FlowHeader: FC = () => {
       })
   }
 
-  const canvasOptions = {
-    backgroundColor,
-    onclone: cloneDoc => {
-      // Add left and right padding on the selected screenshot
-      cloneDoc.getElementById(flow.id).style.padding = '0 12px'
-      cloneDoc
-        .querySelectorAll('[data-download-hide]')
-        .forEach(d => (d.style.display = 'block'))
-    },
-    // Enable map background
-    useCORS: true,
-  }
-
-  const handleDownloadAsPNG = () => {
-    const canvas = document.getElementById(flow.id)
-    import('html2canvas').then((module: any) =>
-      module.default(canvas as HTMLDivElement, canvasOptions).then(result => {
-        downloadImage(result.toDataURL(), `${flow.name}.png`)
-      })
-    )
-  }
-
-  const handleDownloadAsPDF = () => {
-    const canvas = document.getElementById(flow.id)
-    import('html2canvas').then((module: any) =>
-      module.default(canvas as HTMLDivElement, canvasOptions).then(result => {
-        import('jspdf').then((jsPDF: any) => {
-          const doc = new jsPDF.default({
-            orientation: 'landscape',
-            unit: 'pt',
-            format: 'a4',
-          })
-
-          // Background Color
-          doc.setFillColor(backgroundColor)
-          doc.rect(0, 0, 850, 600, 'F')
-
-          const paddingX = 24
-
-          // Pipelist screenshot
-          const imgData = result.toDataURL('image/png')
-
-          // a4 format landscape size in pt unit is 842 x 595
-          const pageWidth = 842
-          const pageHeight = 595
-          const imgWidth = pageWidth - paddingX * 2
-          const imgHeight = (result.height * imgWidth) / result.width
-          let heightLeft = imgHeight
-          let position = 0
-
-          // First page
-          doc.addImage(imgData, 'PNG', paddingX, 0, imgWidth, imgHeight)
-          heightLeft -= pageHeight
-
-          // Add multiple pages
-          while (heightLeft >= 0) {
-            position = heightLeft - imgHeight
-            doc.addPage()
-
-            // Add background color
-            doc.setFillColor(backgroundColor)
-            doc.rect(0, 0, 850, 600, 'F')
-
-            // Add piplist screenshot
-            doc.addImage(
-              imgData,
-              'PNG',
-              paddingX,
-              position,
-              imgWidth,
-              imgHeight
-            )
-            heightLeft -= pageHeight
-          }
-          doc.save(`${flow.name}.pdf`)
-        })
-      })
-    )
-  }
-
   const generateLink = () => {
     event('Show Share Menu', {share: !!share ? 'sharing' : 'not sharing'})
 
@@ -323,91 +159,6 @@ const FlowHeader: FC = () => {
     /* eslint-disable no-console */
     console.log(JSON.stringify(serialize(flow), null, 2))
     /* eslint-enable no-console */
-  }
-
-  const handleClone = async () => {
-    event('clone_notebook')
-    const clonedId = await clone(flow.id)
-    setShare(null)
-    history.push(
-      `/orgs/${orgID}/${PROJECT_NAME_PLURAL.toLowerCase()}/${clonedId}`
-    )
-  }
-
-  const handleDelete = () => {
-    event('delete_notebook')
-    deletePinnedItemByParam(flow.id)
-    remove(flow.id)
-    history.push(`/orgs/${orgID}/${PROJECT_NAME_PLURAL.toLowerCase()}`)
-  }
-
-  const handleViewPublish = () => {
-    event('viewing_publish_history')
-    const [first, second] = versions
-    // accounts for the draft state
-    let versionId = first.id
-    if (first.id === 'draft') {
-      versionId = second.id
-    }
-    history.push(
-      `/orgs/${orgID}/${PROJECT_NAME_PLURAL.toLowerCase()}/${
-        flow.id
-      }/versions/${versionId}`
-    )
-  }
-
-  const menuItems: MenuItemType[] = [
-    {
-      type: 'menuitem',
-      title: 'Clone',
-      onClick: handleClone,
-      icon: IconFont.Duplicate_New,
-    },
-    {
-      type: 'menuitem',
-      title: 'Download as PNG',
-      onClick: handleDownloadAsPNG,
-      icon: IconFont.Download_New,
-    },
-    {
-      type: 'menuitem',
-      title: 'Download as PDF',
-      onClick: handleDownloadAsPDF,
-      icon: IconFont.Download_New,
-    },
-    {
-      type: 'menuitem',
-      title: 'Delete',
-      onClick: handleDelete,
-      icon: IconFont.Trash_New,
-      testID: 'flow-menu-button-delete',
-    },
-  ]
-
-  if (isFlagEnabled('flowPublishLifecycle')) {
-    menuItems.splice(
-      0,
-      0,
-      {
-        type: 'menuitem',
-        title: 'Save to version history',
-        onClick: handlePublish,
-        icon: IconFont.Disks,
-      },
-      {
-        type: 'menuitem',
-        title: 'Version history',
-        onClick: handleViewPublish,
-        icon: IconFont.Layers,
-        disabled: () => {
-          if (versions.length > 1) {
-            return false
-          }
-          return versions[0]?.id === 'draft'
-        },
-      },
-      {title: 'divider', type: 'divider'}
-    )
   }
 
   if (!flow) {
@@ -450,7 +201,7 @@ const FlowHeader: FC = () => {
                   }
                   titleText={`Share ${PROJECT_NAME}`}
                 />
-                <MenuButton menuItems={menuItems} />
+                <MenuButton handleResetShare={() => setShare(null)} />
               </>
             )}
             <FeatureFlag name="flow-snapshot">
